pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mtk"
version = "0.1.0"
description = "Mail Toolkit - Personal email archive management with semantic search and relationship mapping"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "spinoza" }
]
keywords = ["email", "mail", "archive", "notmuch", "search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
notmuch = [
    "notmuch2>=0.38",
]
semantic = [
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
]
mcp = [
    "mcp>=1.0.0",
]
imap = [
    "imapclient>=3.0.0",
    "keyring>=25.0.0",
]
imap-oauth = [
    "google-auth-oauthlib>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
all = [
    "mtk[notmuch,semantic,mcp,imap,dev]",
]

[project.scripts]
mtk = "mtk.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/mtk"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/mtk"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]
